Didac: casteo a entero de la posicon de la camara, avanza pixel a pixel

+++ core/src/es/cdv/GameStage.java
@@ -125,7 +125,7 @@ public class GameStage extends Stage {
                 posX += dist * percent;
 
 
-                getCamera().position.x = posX;
+                getCamera().position.x = ((int) posX);
             }
